SPARK-30004: Allow merge UserDefinedType into normal DataType
In case you write a UDT, you always need to read it with the
UDT registered. In many cases you want to write it, and then
convert it into a native DataType.

In the case of Delta or when appending a partition, you can
write to the same table and then it needs to be able to
convert merge the UDT into the native type again.

* Add a test to the DataTypeSuite.scala