Reduce job failures during decommissioning
This PR reduces the prospect of a job loss during decommissioning. It
fixes two holes in the current decommissioning framework:

- (a) Loss of decommissioned executors is not treated as a job failure:
We know that the decom'd executor would be dying soon, so its death is
clearly not caused by the application.

- (b) Shuffle files on the decommissioned host are cleared when the
first fetch failure is detected from a decommissioned host: This is a
bit tricky in terms of when to clear the shuffle state ? Ideally you
want to clear it the millisecond before the shuffle service on the node
dies (or the executor dies when there is no external shuffle service) --
too soon and it could lead to some wasteage and too late would lead to
fetch failures.

The approach here is to do this clearing when the very first fetch
failure is observed on the decom'd block manager, without waiting for
other blocks to also signal a failure.

(The commits changes for these two features share some common code and
hence I am packing both of them into the same commit.)

Added a new unit test `DecommissionWorkerSuite` to test the new
behavior.

TODO:
- Should I add a feature flag to guard these two behaviors ?
- Should this to be two separate PR's or two commits in the same PR ?