Expose a (protected) /workers/kill endpoint on the MasterWebUI
This allows an external agent to inform the Master that certain
hosts are being decommissioned.

This alternative is suitable for some environments that cannot trigger a
clean up hook on the Worker that is needed today to inform the Master.

This new API also allows the Master to be informed of all hosts being
decommissioned in bulk by specifying a list of hostnames.

This API is merely a new entry point into the existing decommissioning
logic. It does not change how the decommissioning request is handled in
its core.

Added unit tests