Expose a (protected) /workers/kill endpoint on the MasterWebUI
This allows an external agent to inform the Master that certain nodes
(or host-ports) are being decommissioned.

This alternative is suitable for some environments that cannot trigger a
clean up hook on the Worker that is needed today to inform the Master.

This new API also allows the Master to be informed of all nodes being
decommissioned in bulk. We allow specifying just the hostname, in which
case all workers on the host will be decommissioned. Or one can specify
only certain worker host:port's.

This API is merely a new entry point into the existing decommissioning
logic. It does not change how the decommissioning request is handled in
its core.

Added unit tests