Plumb whether a worker would also be decommissioned along with executor
This PR is a giant plumbing PR that plumbs a DecommissionInfo along
with the DecommissionExecutor message. We can embellish DecommissionInfo
in the future for knowing how long the executor has (for spot kills) and
the like.

The motivation is to know whether a decommissioned executor would also
be loosing shuffle files -- and thus it is important to know whether the
worker (node) would also be decommissioned. This is similar to the
existing `workerLost` field in the `SlaveLost` message.